Pinging Implementation
The server regularly sends a message to each client and expects a reply; if no reply is
received, the player is timed out and dropped from their game, if any.

For this I implemented it in client and Player.
What is left is to figure out where to use the method in Client.